+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - add developer page
 - update library page with provider
 - update dev page
+- add new book menu sheet
 
 ## [0.8.1] - 2025-11-20
 - library sorting to case-insensitive