+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - update dev page
 - add new book menu sheet
 - add link to licenses on settings page
+- add mock books to dev page
 
 ## [0.8.1] - 2025-11-20
 - library sorting to case-insensitive