+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.9.0] - open
+## [0.9.0] - 2026-02-19
 - add developer mode setting
 - add developer page
 - update library page with provider